Make use of the new evaluate measure code from at least one single library call.  Fixed invalid error handling before.  All tests pass.
